% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{combine_testingTree}
\alias{combine_testingTree}
\title{combine hic_diff from all batches}
\usage{
combine_testingTree(hicDbList)
}
\arguments{
\item{hicDbList}{a list of \code{treeHiCDataSet} objects resulting from batch mode}
}
\value{
a list (\code{testingTree}) which is a combined result from individual batch's \code{testingTree}.
}
\description{
combine hic_diff from all batches
}
