% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{convertForVtk}
\alias{convertForVtk}
\title{convert rep to for nx3 matrix: f(x,y) | x | y
whether taking lower triangle (default) or not
in this notation: x is index for row, y index for col}
\usage{
convertForVtk(hicMat)
}
\arguments{
\item{hicMat}{a squared HiC matrix}
}
\description{
convert rep to for nx3 matrix: f(x,y) | x | y
whether taking lower triangle (default) or not
in this notation: x is index for row, y index for col
}
