% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_batches}
\alias{create_batches}
\title{set up data for running batch version}
\usage{
create_batches(hicDb, batchSize = 10^6, nBatches = 5)
}
\arguments{
\item{hicDb}{a \code{treeHiCDataSet} object}

\item{batchSize}{number of cells in batches}

\item{nBatches}{number of batches}
}
\value{
batches of csv files and hicDb objects
}
\description{
set up data for running batch version
}
