% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_testing_tree}
\alias{create_testing_tree}
\title{get cell ID (or vertexID vtk pipeline) for pLevel1 and pLevel2 with inputs
from Topological Toolkit (ttk)}
\usage{
create_testing_tree(pathThresholdPLevelList, pathMSComplexPLevel1, d_height)
}
\arguments{
\item{pathThresholdPLevelList}{list of csv files for ThresholdPLevel data}

\item{pathMSComplexPLevel1}{MSComplex partitions from pLevel1}

\item{d_height}{data frame of 3 columns storing xy coords and height/difference values}
}
\value{
slot \code{testingTree} for object \code{treeHiCDataSet} object
}
\description{
get cell ID (or vertexID vtk pipeline) for pLevel1 and pLevel2 with inputs
from Topological Toolkit (ttk)
}
