% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{eval_fdr_table_permutation}
\alias{eval_fdr_table_permutation}
\title{Title}
\usage{
eval_fdr_table_permutation(mat_pvals, i.range, j.range, alpha,
  use_adjusted_pvals)
}
\arguments{
\item{use_adjusted_pvals}{}
}
\description{
Title
}
