% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{evalPvals}
\alias{evalPvals}
\title{add p-values information to
\code{testingTree}}
\usage{
evalPvals(testingTree, mat_pvals)
}
\arguments{
\item{testingTree}{a slot \code{testingTree} of object \code{treeHiCDataSet}}

\item{mat_pvals}{a squared matrix of (raw) p-values to perform testing}
}
\description{
add p-values information to
\code{testingTree}
}
