% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_partitions}
\alias{get_partitions}
\title{generate Morse-Smale partitions}
\usage{
get_partitions(path, pLevelGrid)
}
\arguments{
\item{path}{path to a working folder}

\item{pLevelGrid}{a list whose elements contain different persistence levels of
\code{testingTree}}
}
\description{
generate Morse-Smale partitions
}
