% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{get_perm_pvals}
\alias{get_perm_pvals}
\title{calculate permutation p-values for each cell}
\usage{
get_perm_pvals(mat1, mat2)
}
\arguments{
\item{mat1}{a HiC contact matrix for the first condition}

\item{mat2}{a HiC contact matrix for the first condition}
}
\value{
a squared matrix storing pvalues from permutation
}
\description{
calculate permutation p-values for each cell
}
