% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_persistence_curve}
\alias{get_persistence_curve}
\title{generate persistence curve}
\usage{
get_persistence_curve(path)
}
\arguments{
\item{path}{path to a working folder}
}
\description{
generate persistence curve
}
