% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{hic_diff_simplified}
\alias{hic_diff_simplified}
\title{get cell ID (or vertexID vtk pipeline) for pLevel1 and pLevel2 with inputs
from Topological Toolkit (ttk). This is the simplified version of hic_diff()
where the input of MSComplex is not required (i.e.,\code{eval_MSComplex = F}).}
\usage{
hic_diff_simplified(hicDb, mat_pvals, use_adjusted_pvals = TRUE, alpha,
  batch_mode = FALSE)
}
\arguments{
\item{mat_pvals}{: a squared matrix of (raw) p-values to perform testing}

\item{use_adjusted_pvals}{: Logical, should adjusted p-values be used?}

\item{alpha}{: FDR control at \code{alpha} level}

\item{batch_mode}{logical; TRUE if running on large HiC matrix. Default to FALSE}

\item{object}{: a \code{treeHiCDataSet} object}
}
\value{
add additional slots \code{testingTree, hic_diff_result}
}
\description{
get cell ID (or vertexID vtk pipeline) for pLevel1 and pLevel2 with inputs
from Topological Toolkit (ttk). This is the simplified version of hic_diff()
where the input of MSComplex is not required (i.e.,\code{eval_MSComplex = F}).
}
