% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{HiCDataSetFromMatrix}
\alias{HiCDataSetFromMatrix}
\title{prepare HiC contact matrices for differential interaction. For this function, you should provide
count matrices, or txt files specified in the column information \code{colData}.}
\usage{
HiCDataSetFromMatrix(object, contactMatrixList = list(), colData, path)
}
\arguments{
\item{object}{: a \code{treeHiCDataSet} object}

\item{contactMatrixList}{: a list of HiC contact matrices to perform differential interaction}

\item{colData}{: a table of sample information}

\item{path}{: path to the working folder}
}
\value{
a contact matrix list with respect to the column information \code{colData}
}
\description{
prepare HiC contact matrices for differential interaction. For this function, you should provide
count matrices, or txt files specified in the column information \code{colData}.
}
