% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot_persistence_curve}
\alias{plot_persistence_curve}
\title{plot persistence curve}
\usage{
plot_persistence_curve(pLevelGridList, path)
}
\arguments{
\item{pLevelGridList}{slot pLevelGrid from \code{treeHiCDataSet} object}

\item{path}{path from \code{treeHiCDataSet} object}
}
\description{
plot persistence curve
}
