% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{selectPLevelGrid}
\alias{selectPLevelGrid}
\title{pich the most persistent number of partitions and extremas}
\usage{
selectPLevelGrid(object, maxDepth = 0)
}
\arguments{
\item{object}{: a \code{treeHiCDataSet} object}

\item{maxDepth}{: maximum depth of testing tree}

\item{path}{: path to the working folder}
}
\value{
return the optimal persistent level
}
\description{
pich the most persistent number of partitions and extremas
}
