% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{treeHiCDataSet-class}
\alias{treeHiCDataSet-class}
\title{A class to store HiC data and all related objects for differential interaction analysis.}
\description{
A class to store HiC data and all related objects for differential interaction analysis.
}
\section{Slots}{

\describe{
\item{\code{contactMatrixList}}{A list stores HiC contact matrices}

\item{\code{d_height}}{data frame of 3 columns storing xy coords and height/difference values}

\item{\code{persDiag}}{data frame storing persistence function}

\item{\code{pLevelGrid}}{a list whose elements contain different persistence levels of
\code{testingTree}}

\item{\code{path}}{path to the working folder}

\item{\code{testingTree}}{a list whose elements are nodes of \code{testingTree}}

\item{\code{checkTree}}{a logical list to perform hierarchical testing}

\item{\code{hic_diff_result}}{a matrix containing location of cell being tested with its p-values}
}}

