% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{eval_fdr_table}
\alias{eval_fdr_table}
\title{Eval fdr table for simulation}
\usage{
eval_fdr_table(pvals, hic_dim, i.range.truth, j.range.truth,
  FIND_model = FALSE)
}
\arguments{
\item{FIND_model}{logical; TRUE if evaluating FDR in FIND_model}
}
\description{
Eval fdr table for simulation
}
