% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_diff_hic}
\alias{get_diff_hic}
\title{Title}
\usage{
get_diff_hic(testingTree, alpha, use_adjusted_pvals = TRUE,
  batch_mode = FALSE)
}
\arguments{
\item{testingTree}{a slot \code{testingTree} of object \code{treeHiCDataSet}}

\item{alpha}{FDR control at \code{alpha} level}

\item{use_adjusted_pvals}{Logical, default to TRUE if adjusted pvalues are used}

\item{batch_mode}{logical; TRUE if running on a large HiC matrix. Default to FALSE}
}
\description{
Title
}
